Require microgames to post a 'ready' event.
This breaks compatibility with existing microgames at the postMessage
protocol level, but *not* at the tinygame.js level. Tinygame now
automatically sends a 'ready' event on page load, but the default
behavior can be prevented via Tinygame.loading() and a later
call to Tinygame.loaded().

This gets rid of a lot of complexity involved in the `customloadstart`
and `customloadend` business.